+++ cofi-share/src/main/java/com/omelan/cofi/share/model/Step.kt
@@ -124,7 +124,7 @@ data class Step(
     val name: String,
     val time: Int? = null,
     val type: StepType,
-    val value: Double? = null,
+    val value: Float? = null,
 ) : SharedData {
     override fun serialize(): JSONObject = JSONObject().let {
         it.put(jsonId, id)
@@ -149,8 +149,8 @@ fun JSONObject.getIntOrNull(key: String) = try {
     null
 }
 
-fun JSONObject.getDoubleOrNull(key: String) = try {
-    getDouble(key)
+fun JSONObject.getFloatOrNull(key: String) = try {
+    getDouble(key).toFloat()
 } catch (e: Exception) {
     null
 }
@@ -161,7 +161,7 @@ fun JSONObject.toStep(withId: Boolean = false, recipeId: Long? = null) = if (wit
         name = getString(jsonName),
         recipeId = recipeId?.toInt() ?: getInt(jsonRecipeId),
         time = getIntOrNull(jsonTime),
-        value = getDoubleOrNull(jsonValue),
+        value = getFloatOrNull(jsonValue),
         orderInRecipe = getInt(jsonOrderInRecipe),
         type = StepTypeConverter().stringToStepType(getString(jsonType)),
     )
@@ -170,7 +170,7 @@ fun JSONObject.toStep(withId: Boolean = false, recipeId: Long? = null) = if (wit
         name = getString(jsonName),
         recipeId = recipeId?.toInt() ?: getInt(jsonRecipeId),
         time = getIntOrNull(jsonTime),
-        value = getDoubleOrNull(jsonValue),
+        value = getFloatOrNull(jsonValue),
         orderInRecipe = getInt(jsonOrderInRecipe),
         type = StepTypeConverter().stringToStepType(getString(jsonType)),
     )
